# openapi.yaml
openapi: 3.0.4
info:
  title: Strava API v3
  description: FastAPI implementation of Strava API v3
  version: 3.0.0
paths:
  /athletes/{athlete_id}/stats:
    get:
      tags:
        - Athlete
      summary: Get Athlete Stats
      description: Returns the activity stats of an athlete.
      operationId: getAthleteStats
      parameters:
        - name: athlete_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the athlete
            title: Athlete Id
          description: The identifier of the athlete
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityStats'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /athlete:
    get:
      tags:
        - Athlete
      summary: Get Authenticated Athlete
      description: Returns the currently authenticated athlete.
      operationId: getAuthenticatedAthlete
      parameters:
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedAthlete'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - Athlete
      summary: Update Authenticated Athlete
      description: Update the currently authenticated athlete.
      operationId: updateAuthenticatedAthlete
      parameters:
        - name: weight
          in: query
          required: true
          schema:
            type: number
            description: The weight of the athlete in kilograms
            title: Weight
          description: The weight of the athlete in kilograms
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedAthlete'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /athlete/zones:
    get:
      tags:
        - Athlete
      summary: Get Authenticated Athlete Zones
      description: Returns the authenticated athlete's heart rate and power zones.
      operationId: getAuthenticatedAthleteZones
      parameters:
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /segments/{segment_id}:
    get:
      tags:
        - Athlete
      summary: Get Segment By Id
      description: Returns the specified segment.
      operationId: getSegmentById
      parameters:
        - name: segment_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the segment
            title: Segment Id
          description: The identifier of the segment
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedSegment'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /segments/starred:
    get:
      tags:
        - Athlete
      summary: Get Starred Segments
      description: List of the authenticated athlete's starred segments.
      operationId: getStarredSegments
      parameters:
        - name: page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Page number
            default: 1
            title: Page
          description: Page number
        - name: per_page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Number of items per page
            default: 30
            title: Per Page
          description: Number of items per page
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SummarySegment'
                title: Response Getstarredsegments
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /segments/explore:
    get:
      tags:
        - Athlete
      summary: Explore Segments
      description: Returns the top 10 segments matching a specified query.
      operationId: exploreSegments
      parameters:
        - name: bounds
          in: query
          required: true
          schema:
            type: array
            items:
              type: number
            description: Rectangular boundary for search
            title: Bounds
          description: Rectangular boundary for search
        - name: activity_type
          in: query
          required: false
          schema:
            anyOf:
              - type: string
            nullable: true
            description: Desired activity type
            title: Activity Type
          description: Desired activity type
        - name: min_cat
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 5
                minimum: 0
            nullable: true
            description: Minimum climbing category
            title: Min Cat
          description: Minimum climbing category
        - name: max_cat
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 5
                minimum: 0
            nullable: true
            description: Maximum climbing category
            title: Max Cat
          description: Maximum climbing category
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /segment_efforts:
    get:
      tags:
        - Athlete
      summary: Get Segment Efforts
      description: Returns segment efforts for a given segment.
      operationId: getSegmentEfforts
      parameters:
        - name: segment_id
          in: query
          required: true
          schema:
            type: integer
            description: The identifier of the segment
            title: Segment Id
          description: The identifier of the segment
        - name: start_date_local
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: date-time
            nullable: true
            description: Start date filter
            title: Start Date Local
          description: Start date filter
        - name: end_date_local
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: date-time
            nullable: true
            description: End date filter
            title: End Date Local
          description: End date filter
        - name: per_page
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
            nullable: true
            description: Number of items per page
            default: 30
            title: Per Page
          description: Number of items per page
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /segment_efforts/{effort_id}:
    get:
      tags:
        - Athlete
      summary: Get Segment Effort By Id
      description: Returns a segment effort from an activity.
      operationId: getSegmentEffortById
      parameters:
        - name: effort_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the segment effort
            title: Effort Id
          description: The identifier of the segment effort
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /activities/{activity_id}:
    get:
      tags:
        - Athlete
      summary: Get Activity By Id
      description: Returns the given activity.
      operationId: getActivityById
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the activity
            title: Activity Id
          description: The identifier of the activity
        - name: include_all_efforts
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
            nullable: true
            description: Include all segment efforts
            default: false
            title: Include All Efforts
          description: Include all segment efforts
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedActivity'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /activities/{activity_id}/laps:
    get:
      tags:
        - Athlete
      summary: Get Activity Laps
      description: Returns the laps of an activity.
      operationId: getActivityLaps
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the activity
            title: Activity Id
          description: The identifier of the activity
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /activities/{activity_id}/zones:
    get:
      tags:
        - Athlete
      summary: Get Activity Zones
      description: Returns the zones of a given activity.
      operationId: getActivityZones
      parameters:
        - name: activity_id
          in: path
          required: true
          schema:
            type: integer
            description: The identifier of the activity
            title: Activity Id
          description: The identifier of the activity
        - name: authorization
          in: header
          required: true
          schema:
            type: string
            description: Bearer token for authentication
            title: Authorization
          description: Bearer token for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActivityStats:
      properties:
        recent_ride_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: Recent Ride Totals
        recent_run_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: Recent Run Totals
        ytd_ride_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: Ytd Ride Totals
        ytd_run_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: Ytd Run Totals
        all_ride_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: All Ride Totals
        all_run_totals:
          anyOf:
            - additionalProperties: true
              type: object
          nullable: true
          title: All Run Totals
      type: object
      title: ActivityStats
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError